+++ src/pages/Tasks/styles.js
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  logoOrcStyle: {
+    width: 64,
+    height: 64,
+  },
+
+  addButtonStyle: {
+    height: 42,
+    borderColor: 'transparent',
+    backgroundColor: 'transparent',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 10,
+  },
+
+  addButtonTextStyle: {
+    fontSize: 54,
+    color: '#FFF',
+  },
+});
+
+export default styles;
